Projectile, sound effects, updated scripts
I was able to add projectile based system to add a different kind of layer to the game. Sound effects to the jumping mechanic and the dying mechanic and effects to both running and death. Along with added music and fixing all bugs I might have had in my code.